refactor(profile): /simplify pass on PR peterdrier#376
- UserEmailAuthorizationHandler: move to Authorization/Requirements/ to match
  the Budget/Camp/Team handler convention; replace inline 3-role IsInRole check
  with RoleChecks.IsHumanAdminBoardOrAdmin
- EmailsViewModel: drop IsGoogleServiceEmail (tautology of IsGoogle) and the
  MergePendingEmailIds set (use IsMergePending per row instead)
- UserEmailService.SetPrimaryAsync: short-circuit + only mutate rows whose
  IsPrimary actually changes (mirrors EnsurePrimaryInvariantAsync); a re-click
  on the already-selected radio no longer rewrites every row + evicts the cache
- UserEmailService.AddEmailAsync: drop redundant EnsurePrimaryInvariantAsync
  call (unverified rows can't be primary, so the helper was always a no-op here)
- Strip multi-paragraph PR-history comments in UserEmailService and the
  ProfileController admin-actions banner; keep the WHY trimmed to one line
